pyproject.toml
[project]
name = "ff-srl2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.45",
    "imgui[pyglet]>=2.0.0",
    "matplotlib>=3.9.4",
    "newton-physics",
    "pycollada>=0.9.2",
    "pyglet==2.1.3", # for some reason higher versions dont work for Przemek
    "pyopenhaptics>=1.0.1",
    "trimesh>=4.7.1",
    "usd-core>=25.5.1",
    "warp-lang>=1.9.0",
]


[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com"
explicit = true

[[tool.uv.index]]
url = "https://pypi.org/simple"

[tool.uv.sources]
newton = { path = "../newton", editable = true }
#warp-lang = { path = "../warp", editable = true }
